fixes for brakeman and paginable
