+++ app/models/concerns/date_rangeable.rb
@@ -16,7 +16,8 @@ module DateRangeable
     def by_date_range(field, term)
       date = Date.parse(term) if term[0..1].match(/[0-9]{2}/).present?
       date = Date.parse("1st #{term}") unless date.present?
-      where("#{table_name}.#{field} BETWEEN ? AND ?", date, date.end_of_month)
+      query = "%{table}.%{field} BETWEEN ? AND ?" % { table: table_name, field: field }
+      where(query, date, date.end_of_month)
     end
 
   end
