+++ app/presenters/identifier_presenter.rb
@@ -23,17 +23,6 @@ class IdentifierPresenter
     schemes.select { |scheme| scheme.name.downcase == name.downcase }
   end
 
-  def id_for_display(id:, with_scheme_name: true)
-    return _("None defined") if id.new_record? || id.value.blank?
-
-    without = id.value_without_scheme_prefix
-    prefix = with_scheme_name ? id.identifier_scheme.description + ": " : ""
-    return prefix + id.value unless without != id.value && !without.starts_with?("http")
-
-    "<a href=\"#{id.value}\" class=\"has-new-window-popup-info\"> " \
-      "#{prefix} #{without}</a>"
-  end
-
   private
 
   def load_schemes
